\name{fdesigns-package}
\alias{fdesigns-package}
\alias{fdesigns}
\docType{package}
\title{
\packageTitle{fdesigns}
}
\description{
\packageDescription{fdesigns}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fdesigns}
\packageIndices{fdesigns}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fdesigns}

Maintainer: \packageMaintainer{fdesigns}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
